requirements.txt
appdirs==1.4.4
backcall==0.2.0
black==20.8b1
cbor2==5.2.0
click==7.1.2
dbus-python==1.2.16
decorator==4.4.2
gatt==0.2.7
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
mypy-extensions==0.4.3
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycairo==1.20.0
pydbus==0.6.0
Pygments==2.7.2
PyGObject==3.38.0
regex==2020.10.28
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
pyyaml
paho-mqtt
victron_ble>=0.6.0
